Merged PR 1421: Include Semi-Supervision code
Include Semi-Supervision Code.

**Yae Jee Results**
![image (2).png](https://msktg.visualstudio.com/c507252c-d1be-4d67-a4a1-03b0181c35c7/_apis/git/repositories/0392018c-4507-44bf-97e2-f2bb75d454f1/pullRequests/1421/attachments/image%20%282%29.png)

**FLUTE Results**
![image (5).png](https://msktg.visualstudio.com/c507252c-d1be-4d67-a4a1-03b0181c35c7/_apis/git/repositories/0392018c-4507-44bf-97e2-f2bb75d454f1/pullRequests/1421/attachments/image%20%285%29.png)

The final accuracy of the three models is pretty similar to YJ's results, the only difference is the training loss computation. In her code she is using all the client training data on each round, not only the ones sampled per iteration as in FLUTE. I asked about this to Dimitrios a while ago and he mentioned that this shouldn't be a big issue, just that the loss would oscillate more in our case than in her results, which I believe is the case.

I had to do some slight changes in a couple files inside _/core_ to allow FLUTE to use both aggregation methods: using pseudogradients or state dictionaries, and to pass the iteration round from server to clients during each round, given that YJ's algorithm needed for running _burnout rounds_ before starting to execute the algorithm. Aside from that, there are no major changes.

Let me know if you have any comment!

Pending Tasks
[x] Remove amlt/git files
[x] 3rd Party notice?

Sanity-Checks:
[x] semisupervision:  https://aka.ms/amlt?q=d4h7c
[x] mlm_bert: https://aka.ms/amlt?q=d2p0d
[x] nlg_gru: https://aka.ms/amlt?q=d4fq7